Handle missing registry values.
Warn if AppParameters is missing.

Warn if AppDirectory is missing or unset and choose a fallback directory.
First try to find the parent directory of the application.  If that
fails, eg because the application path is just "notepad" or something,
expand %SYSTEMROOT%.  If THAT fails it's time to give up and go home.

Thanks Arve Knudsen.