Removed run hack.
Previously we used to attempt to run as a service only when the "run"
argument was given, and printed a usage message if no arguments were
provided.

A more appropriate strategy is to check for StartServiceCtrlDispatcher()
returning ERROR_FAILED_SERVICE_CONTROLLER_CONNECT, as that means the
application is not running in a service context.

As a result we no longer write the undocumented "run" argument to the
registry when installing a service.  Technically this means that we
sacrifice backward compatibility with older versions.