Windows 2000 optimisation.
Windows 2000 takes a few seconds to time out when trying and failing to
connect to the service manager if NSSM was run with no arguments.

Check for stdin first.  It won't exist if we are running in a service
context so we can skip the connection attempt unless we're unsure.