Log exit messages more intuitively.
In the case where the application crashed and NSSM then cleaned up its
child processes the events logged were confusing.  First we'd log that
we were going to kill the process tree with exit code X then we'd log
that the service had ended with exit code X.

Logging that the application exited before doing the process tree kill
makes it clearer that the cleanup is in response to the application
exit.