Updated nssm.dsp with missing dependencies.
Opening nssm.dsp in Visual Studio 2010 and accepting conversion to the
latest project format produced a project which would not build because
io.cpp, io.h, process.cpp, process.h and shlwapi.lib had not been added
to nssm.dsp.

They were present in nssm.vcproj from Visual Studio 2008 and it was
possible to load that project and convert to a format which worked.

Thanks Guy Brousseau.