Force UTF-16 output.
If we were compiled as a Unicode application, set stdout and stderr
output mode to UTF-16 text.  Doing so allows French and Italian accented
characters to show up correctly on the console.

We currently have no analogous method for fixing the output of NSSM when
compiled as an ANSI application.