Write canonical AppExit string.
If the user ran "nssm set <servicename> AppExit Default exit" the
default value of AppExit would be set verbatim to "exit" which is in
contrast the documented value "Exit" with normal capitalisation.

There's no harm to that fact, as AppExit is retrieved case-insensitively
but it was nonetheless not working as intended.