Allocate new console by default.
Since the whole purpose of NSSM is to be able to make services out of
applications which expect to have a standard desktop environment, let's
simply allocate a new console for each one rather than do so only when
we try to redirect I/O.

A funky ASCII art banner identifies the console as belonging to the
service.  NSSM itself will detach from the console as soon as possible,
so if the service runs a graphical application the console may never be
seen.

The console can be suppressed by setting the REG_DWORD value
AppNoConsole to 1.

Note that services running console applications on Windows 2000 may
fail if the console is disabled.