Ensure we have a valid service exit time.
Sometimes GetProcessTimes() will succeed but return 0 for the process
exit time, leading check_parent() to conclude that the service had
exited before a child process had started, and thus that the child
process was not part of the service's tree.

We now use initialise the service exit time to the current time before
attempting to get the real exit time, and only override the value if
it's valid.

Thanks Czenda Czendov.