Kill process tree from the top down.
On Windows 8.1 (and maybe other versions), attempting to kill
conhost.exe before its parent cmd.exe will cause the attempt to kill the
latter to fail and the console window to remain active after the service
ends.

We now kill the members of the process tree in reverse order, so the
problem doesn't occur.

Thanks Czenda Czendov.