Handle well-known account names.
Certain well-known accounts can be used to start a service without
entering a password.

"LocalSystem" is the default value of ObjectName in the registry.  It is
not a real account but is an alias for the "NT Authority\System"
account.

"NT Authority\Local Service" has minimal local and no network
privileges.

"NT Authority\Network Service" has minimal local and network privileges.

NSSM now recognises these well-known accounts and will accept them as
ObjectName parameters without complaining that no password was enter.
With the exception of LocalSystem, each can be specified without the "NT
Authority" pseudo-domain prefix.