UAC awareness.
Relaunch NSSM with administrator privileges if an attempt to install,
edit or remove a service fails.

The new process won't inherit the original console so its output won't
be visible.  For that reason we don't attempt to elevate for operations
which need to be able to display messages, such as controlling a service
or editing parameters.

Thanks to various people for requesting the feature but mainly to Marco
Certelli for advice on the implementation.