Really get canonical name in open_service().
The open_service() function would return the canonical service name if
passed a display name but would return the requested name as-is when
called with a valid service name.  We now retrieve the display name and
map back to the key name so that the returned string is capitalised
exactly as it appears in the services database.