Allow editing service dependencies.
Service dependencies can now be queried or set on the command line via
the DependOnService and DependOnGroup parameters, or via the GUI in the
new Dependencies tab.

Service dependencies can be set using either their key name or display
name.  For example:

    nssm set <servicename> DependOnService RpcSs

which is equivalent to:

    nssm set <servicename> DependOnService "Remote Procedure Call (RPC)"

Group dependencies can be set with or without the group identifier
prefix, which for some reason is not actually described in the
documentation; instead readers are invited to consult the header files
for the value of SC_GROUP_IDENTIFIER, which it turns out is the +
symbol.  For example:

    nssm set <servicename> DependOnGroup +UIGroup

which is equivalent to:

    nssm set <servicename> DependOnGroup UIGroup

The prefix is always printed when query group dependencies.

The GUI presents a editbox into which either service or group
dependencies can be typed.  The group prefix is mandatory when editing
dependencies via the GUI.