Help git figure out how to diff .mc and .rc files.
Visual Studio wants .mc and .rc files to be UTF-16LE, which git thinks
are binary.  Use .gitattributes to force treating them as text only when
diffing, explicitly not handling them as text when submitting as they
must retain their encoding to keep Visual Studio happy.

Thanks to Mathias Breiner for proposing the use of .gitattributes in the
repo.  I had the rules in .git/info/exclude in my working copy for a
long time and forgot about the issue!